Added mpileup and varscan
Used commandline instead of pysam and pybedtools, since I was not able
to get bamfile output from pybedtools intersect (asked forum). Next I
was not able to easily reproduce the pileup output format in pysam.
This was easiest and fastest way. I will try to do a improvement later
on.